src/rkstr8/data/lambda/job_submitter_requirements.txt
arnparse==0.0.2
awscli==1.27.51
bitmath==1.3.3.1
boto==2.49.0
boto3==1.26.51
botocore==1.29.51
colorama==0.4.4
docutils==0.16
jmespath==1.0.1
networkx==3.0
pyasn1==0.4.8
python-dateutil==2.8.2
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
smart-open==6.3.0
urllib3==1.26.14